ANDROID: rust_binder: use u64 for death cookie
The C Binder driver uses binder_uintptr_t for the cookie, which always
corresponds to a 64-bit integer.

Bug: 430659969

+++ kernel_platform/common/drivers/android/binder/node.rs
@@ -882,7 +882,7 @@ struct NodeDeathInner {
 pub(crate) struct NodeDeath {
     node: DArc<Node>,
     process: Arc<Process>,
-    pub(crate) cookie: usize,
+    pub(crate) cookie: u64,
     #[pin]
     links_track: AtomicTracker<0>,
     /// Used by the owner `Node` to store a list of registered death notifications.
@@ -911,7 +911,7 @@ impl NodeDeath {
     pub(crate) fn new(
         node: DArc<Node>,
         process: Arc<Process>,
-        cookie: usize,
+        cookie: u64,
     ) -> impl PinInit<DTRWrap<Self>> {
         DTRWrap::new(pin_init!(
             Self {
